Estimate the parameters and empirical efficient influence functions

estimate_univariate_cates() estimates the slope of the univariate conditional average treatment effect's linear approximation for continuous and binary outcomes. First, the data is pre-processed and split into v_folds folds. Then, the conditional outcome regression over all the training sets are estimated using the super_learner. The estimated outcome regressions are then used to estimate the expected difference in potential outcomes across treatment conditions for each individual in their respective validation set. The variable importance parameters are then estimated in the validation sets, along with their efficient influence functions.

estimate_univariate_cates(
  data,
  outcome,
  treatment,
  biomarkers,
  super_learner,
  propensity_score_ls,
  v_folds,
  parallel
)

Arguments

data

A tibble object containing the outcome variable, treatment indicator, and covariates. The treatment variable is a factor whose levels correspond to the names of the propensity_score_ls argument.

outcome

A character defining the name of the outcome variable in data. The outcome must be a continuous or a binary factor variable.

treatment

A character indicating the name of the binary treatment variable in data.

biomarkers

A character vector listing the pre-treatment biomarkers variables in data. biomarkers must be a subset of covariates.

super_learner

A Lrnr_sl object used to estimate the conditional outcome regression. If set to NULL, the default, an elastic net regression is used instead. It is best to use this default behaviour when analyzing small datasets.

propensity_score_ls

A named numeric list providing the propensity scores for the treatment conditions. The first element of the list should correspond to the "treatment" condition, and the second to the "control" condition, whatever their names may be.

v_folds

A numeric indicating the number of folds used for V-fold cross-validation. Defaults to 5.

parallel

A logical determining whether to use origami's built-in parallelized cross-validation routines. This parallelization framework is built upon the future suite. Defaults to FALSE.

Value

A list containing two tibble objects. The first, betas_df, contains the estimated beta coefficients for each biomarker across all validation sets. The second, ic_df, is made up of the cross-validated empirical efficient influence functions of every observation in data.