Estimate the Univariate Survival CATEs

estimate_univariate_survival_cates() estimates the slope of the univariate conditional average treatment effect's linear approximation for right-censored time-to-event outcomes. First, the data is pre-processed and split into v_folds folds. Then, the conditional survival and conditional censoring functions are estimated using their respective hazard functions' SuperLearners. These estimates are then used to estimate the expected difference in potential outcomes across treatment conditions for each individual in their respective validation set. The variable importance parameters are then estimated in the validation sets, along with their efficient influence functions.

estimate_univariate_s_cates(
  long_data,
  event,
  censor,
  treatment,
  biomarkers,
  cond_surv_haz_super_learner,
  cond_censor_haz_super_learner,
  propensity_score_ls,
  v_folds,
  parallel
)

Arguments

long_data

A tibble object containing the longitudinal data output by prep_long_data().

event

A character defining the name of the binary variable in the data argument that indicates whether an event occurred. Observations can have an event or be censored, but not both.

censor

A character defining the name of the binary variable in the data argument that indicates a right-censoring event. Observations can have an event or be censored, but not both.

treatment

A character indicating the name of the binary treatment variable in data.

biomarkers

A character vector listing the biomarkers of interest in data.

cond_surv_haz_super_learner

A Lrnr_sl object used to estimate the conditional event hazard model. If set to NULL, the default, an elastic net regression is used instead. It is best to use this default behaviour when analyzing small datasets.

cond_censor_haz_super_learner

A Lrnr_sl object used to estimate the conditional censoring hazard model. If set to NULL, the default, an elastic net regression is used instead. It is best to use this default behaviour when analyzing small datasets.

propensity_score_ls

A named numeric list providing the propensity scores for the treatment conditions. The first element of the list should correspond to the "treatment" condition, and the second to the "control" condition, whatever their names may be.

v_folds

A numeric indicating the number of folds used for V-fold cross-validation. Defaults to 5.

parallel

A logical determining whether to use origami's built-in parallelized cross-validation routines. This parallelization framework is built upon the future suite. Defaults to FALSE.

Value

A list containing two tibble objects. The first, betas_df, contains the estimated beta coefficients for each biomarker across all validation sets. The second, ic_df, is made up of the cross-validated empirical efficient influence functions of every observation in data.