Compute Validation Set Objects

hold_out_calculation() computes all the objects required to estimate the variable importance parameter and compute their empirical efficient influence curve over the validation set, assuming continuous or binary outcomes. It begins by estimated the conditional outcome regression on the training set using the super_learner. Next, the difference in potential outcomes is predicted for the observations in the validation set. Finally, these predicted outcomes are used to estimate the variable importance parameters and efficient influence functions.

hold_out_calculation(
  fold,
  data,
  outcome,
  treatment,
  biomarkers,
  super_learner,
  propensity_score_ls,
  outcome_type
)

Arguments

fold

A fold object (from make_folds()).

data

A tibble object containing the outcome variable, treatment indicator, and covariates. The treatment variable is a factor whose levels correspond to the names of the propensity_score_ls argument.

outcome

A character defining the name of the outcome variable in data. The outcome must be a continuous or a binary factor variable.

treatment

A character indicating the name of the binary treatment variable in data.

biomarkers

A character vector listing the pre-treatment biomarkers variables in data. biomarkers must be a subset of covariates.

super_learner

A Lrnr_sl object used to estimate the conditional outcome regression. If set to NULL, the default, an elastic net regression is used instead. It is best to use this default behaviour when analyzing small datasets.

propensity_score_ls

A named numeric list providing the propensity scores for the treatment conditions. The first element of the list should correspond to the "treatment" condition, and the second to the "control" condition, whatever their names may be.

outcome_type

A character indicating the type of outcome. Currently supported outcomes are "continuous" and "binomial". Here, "binomial" is used for binary outcomes.

Value

A list made up of two objects. The first is the numericvector of biomarker variable importance estimates. The second is the tibble of the empirical efficient influence functions for each biomarker.