Compute Validation Set Objects

hold_out_calculation_survival() computes all the objects required to estimate the univariate CATEs for survival outcomes over the validation set. It begins by estimating the conditional survival and censoring hazard functions on the training set using the cond_surv_haz_super_learner and cond_censor_haz_super_learner, respectively. Next, the difference in survival probabilities is predicted for the observations in the validation set. Finally, these predicted outcomes are used to estimate the variable impotance parameters of each biomarker. Their respective empirical efficient influence functions are computed as well.

hold_out_calculation_survival(
  fold,
  long_data,
  event,
  censor,
  treatment,
  biomarkers,
  cond_surv_haz_super_learner,
  cond_censor_haz_super_learner,
  propensity_score_ls
)

Arguments

fold

A fold object (from make_folds()).

long_data

A tibble object containing the longitudinal data output by prep_long_data().

event

A character defining the name of the binary variable in the data argument that indicates whether an event occurred. Observations can have an event or be censored, but not both.

censor

A character defining the name of the binary variable in the data argument that indicates a right-censoring event. Observations can have an event or be censored, but not both.

treatment

A character indicating the name of the binary treatment variable in data.

biomarkers

A character vector listing the biomarkers of interest in data.

cond_surv_haz_super_learner

A Lrnr_sl object used to estimate the conditional event hazard model. If set to NULL, the default, an elastic net regression is used instead. It is best to use this default behaviour when analyzing small datasets.

cond_censor_haz_super_learner

A Lrnr_sl object used to estimate the conditional censoring hazard model. If set to NULL, the default, an elastic net regression is used instead. It is best to use this default behaviour when analyzing small datasets.

propensity_score_ls

A named numeric list providing the propensity scores for the treatment conditions. The first element of the list should correspond to the "treatment" condition, and the second to the "control" condition, whatever their names may be.

Value

A list made up of two objects. The first is the numericvector of biomarker variable importance estimates. The second is the tibble of the empirical efficient influence functions for each biomarker.