Univariate Conditional Average Treatment Effect Estimation

unicate() performs inference about biomarkers' capacity to modify treatment effects in randomized control trials with continuous or binary outcomes. The strength of each biomarkers' treatment effect modification is captured by an unknown variable importance parameter defined as the slope of the univariate conditional average treatment effect's linear approximation. In all but pathological cases, the larger the absolute value of this parameter, the greater the treatment effect modification. unicate() implements assumption-lean, cross-validated inference procedures about these variable importance parameters based on semiparametric theory. Assuming that the biomarkers have non-zero variance, estimates are generated by an unbiased and consistent estimator. Tests assessing whether these slope parameters are significantly different from zero are also performed; they are valid under the same bounded-variance condition.

unicate(
  data,
  outcome,
  treatment,
  covariates,
  biomarkers,
  super_learner = NULL,
  propensity_score_ls,
  v_folds = 5L,
  parallel = FALSE
)

Arguments

data

A data.frame or tibble object containing the outcome variable, treatment indicator, and pre-treatment covariates. Note that the biomarkers are considered be a subset of the covariates.

outcome

A character defining the name of the outcome variable in data. The outcome must be a continuous or a binary factor variable.

treatment

A character indicating the name of the binary treatment variable in data.

covariates

A character vector listing the pre-treatment covariates variables in data.

biomarkers

A character vector listing the pre-treatment biomarkers variables in data. biomarkers must be a subset of covariates.

super_learner

A Lrnr_sl object. If set to NULL, a default SuperLearner is used. If the outcome variable is continuous, the default's library of base learners is made up of a linear model, penalized linear models (LASSO and elasticnet), a spline regression, XGBoost, a Random Forest, and the mean model. When the outcome variable is binary, the base learner library consists of (penalized) logistic regression models, XGBoost, a Random Forests, and the mean model. The type of outcome is automatically detected.

propensity_score_ls

A named numeric list providing the propensity scores for the treatment conditions. The first element of the list should correspond to the "treatment" condition, and the second to the "control" condition, whatever their names may be.

v_folds

A numeric indicating the number of folds used for V-fold cross-validation. Defaults to 5.

parallel

A logical determining whether to use origami's built-in parallelized cross-validation routines. This parallelization framework is built upon the future suite. Defaults to FALSE.

Value

A tibble with rows corresponding to the specified biomarkers. Each row contains an estimate of the treatment-modification variable importance parameter, its standard error, z-score, and the nominal and adjusted p-values of the accompanying test. FDR and FWER adjustments are performed using the Benjamini-Hochberg method and Holm's procedure, respectively. The biomarkers are ordered by significance.